Refactor sandbox service for better encapsulation
